assignment3/requirements.txt
pandas
numpy
sqlalchemy
requests
psycopg2-binary
python-dotenv